+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 - added `CopyToClipboard.svelte`
 - added `ScrollbarStyler.svelte`
 - added `PLATFORM` and `isPlatformIn(oses, browsers)`
+- `Link`:
+	- spacebar scrolling disabled when focused
+	- added `outlineColor`, `outlineStyle` and `outlineWidth` theme props
+	- added `ariaDescribedBy` and `ariaLabel` props
 
 ## `@svizzle/site` v0.4.1 (next)